Wrap Watch3DViewModel creation in a try catch...
If creation fails, log it, and create a Watch3DViewModelBase instead.
Move many properties from HelixWatch3DViewModel to Watch3DViewModelBase to enable minimal binding.
