+++ src/DynamoCoreWpf/ViewModels/Core/DynamoViewModel.cs
@@ -385,15 +385,7 @@ namespace Dynamo.ViewModels
             }
         }
 
-        public HelixWatch3DViewModel BackgroundPreviewViewModel
-        {
-            get
-            {
-                var result = Watch3DViewModels.FirstOrDefault(vm => vm is HelixWatch3DViewModel);
-
-                return (HelixWatch3DViewModel) result;
-            }
-        }
+        public Watch3DViewModelBase BackgroundPreviewViewModel { get; private set; }
 
         public bool BackgroundPreviewActive
         {
@@ -479,12 +471,25 @@ namespace Dynamo.ViewModels
 
             RenderPackageFactoryViewModel = new RenderPackageFactoryViewModel(Model.PreferenceSettings);
             RenderPackageFactoryViewModel.PropertyChanged += RenderPackageFactoryViewModel_PropertyChanged;
-            if (DynamoModel.IsTestMode) return;
+            
             var backgroundPreviewParams = new Watch3DViewModelStartupParams(Model, this, Resources.BackgroundPreviewName);
 
-            var watch3DViewModel = HelixWatch3DViewModel.Start(backgroundPreviewParams);
+            Watch3DViewModelBase watch3DViewModel;
+            try
+            {
+                watch3DViewModel = HelixWatch3DViewModel.Start(backgroundPreviewParams);
+            }
+            catch (Exception ex)
+            {
+                Model.Logger.Log(ex.Message);
+                Model.Logger.Log("Failed to create Watch3DViewModel. Creating base view model.");
+
+                watch3DViewModel = Watch3DViewModelBase.Start(backgroundPreviewParams);
+            }
+
+            BackgroundPreviewViewModel = watch3DViewModel;
             Watch3DViewModels.Add(watch3DViewModel);
-            watch3DViewModel.PropertyChanged += HelixWatch3DViewModelPropertyChanged;
+            watch3DViewModel.PropertyChanged += Watch3DViewModelPropertyChanged;
         }
 
         private void RenderPackageFactoryViewModel_PropertyChanged(object sender, PropertyChangedEventArgs e)
@@ -500,16 +505,15 @@ namespace Dynamo.ViewModels
             }
         }
 
-        void HelixWatch3DViewModelPropertyChanged(object sender, PropertyChangedEventArgs e)
+        void Watch3DViewModelPropertyChanged(object sender, PropertyChangedEventArgs e)
         {
             switch (e.PropertyName)
             {
                 case "Active":
-                    RaisePropertyChanged("BackgroundPreviewActive");
-                    
+                    RaisePropertyChanged("BackgroundPreviewActive");                 
                     break;
                 case "CanNavigateBackground":
-                    if (!((HelixWatch3DViewModel)BackgroundPreviewViewModel).CanNavigateBackground)
+                    if (!BackgroundPreviewViewModel.CanNavigateBackground)
                     {
                         // Return focus back to Search View (Search Field)
                         SearchViewModel.OnRequestReturnFocusToSearch(this, new EventArgs());
@@ -1467,6 +1471,7 @@ namespace Dynamo.ViewModels
 
         public void ToggleFullscreenWatchShowing(object parameter)
         {
+            if (BackgroundPreviewViewModel == null) return;
             BackgroundPreviewViewModel.Active = !BackgroundPreviewViewModel.Active;
         }
 
@@ -1863,7 +1868,8 @@ namespace Dynamo.ViewModels
 
         internal void ZoomIn(object parameter)
         {
-            if (BackgroundPreviewViewModel.CanNavigateBackground)
+            if (BackgroundPreviewViewModel != null && 
+                BackgroundPreviewViewModel.CanNavigateBackground)
             {
                 var op = ViewOperationEventArgs.Operation.ZoomIn;
                 OnRequestViewOperation(new ViewOperationEventArgs(op));
@@ -1881,7 +1887,8 @@ namespace Dynamo.ViewModels
 
         private void ZoomOut(object parameter)
         {
-            if (BackgroundPreviewViewModel.CanNavigateBackground)
+            if (BackgroundPreviewViewModel != null && 
+                BackgroundPreviewViewModel.CanNavigateBackground)
             {
                 var op = ViewOperationEventArgs.Operation.ZoomOut;
                 OnRequestViewOperation(new ViewOperationEventArgs(op));
