+++ src/DynamoCoreWpf/ViewModels/Watch3D/Watch3DViewModelBase.cs
@@ -1,5 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Collections.Specialized;
 using System.ComponentModel;
 using System.Linq;
@@ -11,6 +10,8 @@ using Dynamo.Core.Threading;
 using Dynamo.Interfaces;
 using Dynamo.Models;
 using Dynamo.Selection;
+using Dynamo.Services;
+using Dynamo.UI.Commands;
 using Dynamo.ViewModels;
 
 namespace Dynamo.Wpf.ViewModels.Watch3D
@@ -104,11 +105,61 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
             get { return true; }
         }
 
+        private bool canNavigateBackground = false;
+        public bool CanNavigateBackground
+        {
+            get
+            {
+                return canNavigateBackground || navigationKeyIsDown;
+            }
+            set
+            {
+                canNavigateBackground = value;
+                RaisePropertyChanged("CanNavigateBackground");
+            }
+        }
+
+        private bool navigationKeyIsDown = false;
+        public bool NavigationKeyIsDown
+        {
+            get { return navigationKeyIsDown; }
+            set
+            {
+                if (navigationKeyIsDown == value) return;
+
+                navigationKeyIsDown = value;
+                RaisePropertyChanged("NavigationKeyIsDown");
+                RaisePropertyChanged("CanNavigateBackground");
+            }
+        }
+
+        public DelegateCommand TogglePanCommand { get; set; }
+
+        public DelegateCommand ToggleOrbitCommand { get; set; }
+
+        public DelegateCommand ToggleCanNavigateBackgroundCommand { get; set; }
+
         internal WorkspaceViewModel CurrentSpaceViewModel
         {
             get
             {
                 return viewModel.Workspaces.FirstOrDefault(vm => vm.Model == model.CurrentWorkspace);
+            }
+        }
+
+        public bool IsPanning
+        {
+            get
+            {
+                return CurrentSpaceViewModel != null && CurrentSpaceViewModel.IsPanning;
+            }
+        }
+
+        public bool IsOrbiting
+        {
+            get
+            {
+                return CurrentSpaceViewModel != null && CurrentSpaceViewModel.IsOrbiting;
             }
         }
 
@@ -128,6 +179,17 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
             logger = parameters.Logger;
 
             RegisterEventHandlers();
+
+            TogglePanCommand = new DelegateCommand(TogglePan, CanTogglePan);
+            ToggleOrbitCommand = new DelegateCommand(ToggleOrbit, CanToggleOrbit);
+            ToggleCanNavigateBackgroundCommand = new DelegateCommand(ToggleCanNavigateBackground, CanToggleCanNavigateBackground);
+        }
+
+        public static Watch3DViewModelBase Start(Watch3DViewModelStartupParams parameters)
+        {
+            var vm = new Watch3DViewModelBase(parameters);
+            vm.OnStartup();
+            return vm;
         }
 
         protected virtual void OnStartup()
@@ -150,8 +212,15 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
             // Override in inherited classes.
         }
 
-        protected virtual void OnActiveStateChanged()
+        private void OnActiveStateChanged()
         {
+            preferences.IsBackgroundPreviewActive = active;
+
+            if (active == false && CanNavigateBackground)
+            {
+                CanNavigateBackground = false;
+            }
+
             UnregisterEventHandlers();
             OnClear();
         }
@@ -395,9 +464,57 @@ namespace Dynamo.Wpf.ViewModels.Watch3D
             // Override in derived classes
         }
 
-        internal virtual void TogglePan(object parameter)
+        #region command methods
+
+        internal void TogglePan(object parameter)
         {
-            // Override in derived classes.
+            CurrentSpaceViewModel.RequestTogglePanMode();
+
+            // Since panning and orbiting modes are exclusive from one another,
+            // turning one on may turn the other off. This is the reason we must
+            // raise property change for both at the same time to update visual.
+            RaisePropertyChanged("IsPanning");
+            RaisePropertyChanged("IsOrbiting");
+            RaisePropertyChanged("LeftClickCommand");
         }
+
+        private static bool CanTogglePan(object parameter)
+        {
+            return true;
+        }
+
+        private void ToggleOrbit(object parameter)
+        {
+            CurrentSpaceViewModel.RequestToggleOrbitMode();
+
+            // Since panning and orbiting modes are exclusive from one another,
+            // turning one on may turn the other off. This is the reason we must
+            // raise property change for both at the same time to update visual.
+            RaisePropertyChanged("IsPanning");
+            RaisePropertyChanged("IsOrbiting");
+            RaisePropertyChanged("LeftClickCommand");
+        }
+
+        private static bool CanToggleOrbit(object parameter)
+        {
+            return true;
+        }
+
+        private void ToggleCanNavigateBackground(object parameter)
+        {
+            if (!Active)
+                return;
+
+            CanNavigateBackground = !CanNavigateBackground;
+
+            InstrumentationLogger.LogAnonymousScreen(CanNavigateBackground ? "Geometry" : "Nodes");
+        }
+
+        private bool CanToggleCanNavigateBackground(object parameter)
+        {
+            return true;
+        }
+
+        #endregion
     }
 }
